refactor: reorganize imports and rename hooks for clarity
