+++ src/api/mutations/index.ts
@@ -1,7 +1,3 @@
 export * from './useAcceptFriendsRequests';
-export * from './useBlockFriend';
-export * from './useCreateFriendsRequests';
 export * from './useDeleteAvatar';
-export * from './useMailVerification';
 export * from './useRejectFriendsRequests';
-export * from './useUnblockFriend';
