+++ src/app/(private)/friend/_hooks/useUnblockFriend.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import type { HttpResponse } from '~/api';
-import { fetcher } from '../base';
-import { queryKeys } from '../queryKey';
+import { fetcher } from '~/api/base';
+import { queryKeys } from '~/api/queryKey';
 
 interface Request {
   friendId: number;
